+++ src/Helpers/EspionageHelper.php
@@ -156,10 +156,10 @@ class EspionageHelper
             ],
             [
                 'name' => 'Sabotage Boats',
-                'description' => 'Destroys 2.3% boats',
+                'description' => 'Destroys 2.5% boats',
                 'key' => 'sabotage_boats',
                 'attrs' => ['resource_boats'],
-                'percentage' => 2.3,
+                'percentage' => 2.5,
             ],
             [
                 'name' => 'Incite Chaos',